% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_cor.R
\name{eda_double}
\alias{eda_double}
\title{Calculate EDA Correlation Results}
\usage{
eda_double(
  series1,
  series2,
  m = NULL,
  s = NULL,
  alternative = c("two.sided", "greater", "less")
)
}
\arguments{
\item{series1}{Numerical vector containing NA for missing values.}

\item{series2}{Numerical vector containing NA for missing values with the same length of series1.}

\item{m}{Optional, scaling factor (autocalculated).}

\item{s}{Optional, cutoff series length (autocalculated).}

\item{alternative}{Optional, "two.sided" (Default), "greater", "less".}
}
\value{
A result DataFrame to pass to plotting_double().
}
\description{
Calculate EDA Correlation Results
}
