% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_double.R
\name{eda_plotd}
\alias{eda_plotd}
\title{Create EDA Correlation Plots}
\usage{
eda_plotd(results, scale = NULL, output_file = NULL)
}
\arguments{
\item{results}{Data.frame containing correlation analysis results generated by eda_cor().}

\item{scale}{Numeric, scaling factor for point sizes, autoset by default.}

\item{output_file}{Optional, filename to save plot (NULL to skip saving).}
}
\value{
A gridExtra arranged plot object.
}
\description{
Create EDA Correlation Plots
}
