% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_single.R
\name{eda_plots}
\alias{eda_plots}
\title{Create EDA Analysis Plots}
\usage{
eda_plots(results, scale = NULL, output_file = NULL)
}
\arguments{
\item{results}{Data.frame from eda_analyze().}

\item{scale}{Numeric, scaling factor for point sizes, autoset by default.}

\item{output_file}{Optional, filename to save plot (NULL to skip saving).}
}
\value{
A gridExtra arranged plot object.
}
\description{
Create EDA Analysis Plots
}
