% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_ana.R
\name{eda_single}
\alias{eda_single}
\title{Perform Exploratory Data Analysis on Time Series}
\usage{
eda_single(series, m = NULL, s = NULL)
}
\arguments{
\item{`series`}{A \verb{numerical vector} containing NA for missing values.}

\item{`m`}{A \code{numeric} positive integer, subsampling parameters. Optional
(default: autocalculated).}

\item{`s`}{A \code{numeric} ...}
}
\value{
A result \code{data.frame} to pass to plotting_single().
}
\description{
A short description...
}
\examples{
 \dontrun{
output <- eda_single(
  series = c(1.2, 3.4, 2.5, 4.1, 5.6),
  m = 1,
  s = 6
)
}
}
\author{
Vladimiro Boselli, phD (2025) \email{boselli.v@irea.cnr.it}
}
