% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{read_ts_data}
\alias{read_ts_data}
\title{Read Time Series Data from File}
\usage{
read_ts_data(file_path, ...)
}
\arguments{
\item{file_path}{Path to the input file}

\item{...}{Additional arguments passed to `scan()`}
}
\value{
Numeric vector with the data
}
\description{
Read Time Series Data from File
}
