% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_coupled.R
\name{TH_coupled}
\alias{TH_coupled}
\title{Perform hierarchical moving-window statistical analysis for time series.}
\usage{
TH_coupled(
  series1,
  series2,
  m = NULL,
  s = NULL,
  alpha = 0.1,
  mode = "all",
  alternative = c("two.sided", "greater", "less")
)
}
\arguments{
\item{alternative}{Optional, "two.sided" (Default), "greater", "less".}

\item{`series1`}{a \verb{numerical vector} containing NA for missing values.}

\item{`series2`}{a \verb{numerical vector} containing NA for missing values. It
must be of the same length of \code{series1}.}

\item{`m`}{a \code{numeric} positive integer, subsampling parameter. Optional
(default: autocalculated).}

\item{`s`}{a \code{numeric} positive integer, cutoff parameters for subseries
legth. Optional (default: autocalculated).}

\item{`mode`}{a \code{character} string specifying the computation mode. One of:
"all", "pearson", "kendall", "both", "pearson_with_p", "kendall_with_p".}

\item{`alpha`}{a \code{numeric} significance level for flagging (default: 0.1).}
}
\value{
A result \code{data.frame} to pass to TH_plotc() with attributes.
}
\description{
Perform hierarchical moving-window statistical analysis for coupled time
series. It returns analysis on moving correlations using Pearson's
correlation coefficient and Mann-Kendall correlation coefficient. The results
are in a \code{data.frame} to be passed to TH_plotc() function.
}
\examples{
 \dontrun{
output <- TH_coupled(
  series1 = rnorm(200),
  series2 = rnorm(200),
  m = 2,
  s = 6,
  alpha = 0.05,
  mode = "pearson_with_p",
  alternative = "less"
)
}
}
\author{
Vladimiro Andrea Boselli, (2025) \email{boselli.v@irea.cnr.it}
}
