% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_MK_Corr.R
\name{TH_MK_Corr}
\alias{TH_MK_Corr}
\title{Mann-Kendall Tau Test with Knight's Algorithm}
\usage{
TH_MK_Corr(x, y, alternative = c("two.sided", "greater", "less"))
}
\arguments{
\item{`x`}{a \verb{numerical vector}.}

\item{`alternative`}{Type of alternative hypothesis ("two.sided", "greater", "less")}
}
\value{
A \code{htest} object with results
\item{statistic}{Mann-Kendall tau statistic}
\item{p.value}{p-value}
\item{alt}{Alternative Hypotesis}
}
\description{
Efficient implementation of the Mann-Kendall test for correlation using
Knight's algorithm of (O(n log n) complexity).
}
\examples{
 \dontrun{
}
}
\author{
Vladimiro Andrea Boselli, (2025) \email{boselli.v@irea.cnr.it}
}
