% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_MK_Trend.R
\name{TH_MK_Trend}
\alias{TH_MK_Trend}
\title{Mann-Kendall Trend Test with Knight's Algorithm}
\usage{
TH_MK_Trend(data_series)
}
\arguments{
\item{`data_series`}{a \verb{numerical vector} containing NA for missing values.}
}
\value{
A list with components:
\item{S}{Mann-Kendall S statistic}
\item{Var.S}{Variance of S}
\item{Z}{Z-score}
\item{p.value}{Two-sided p-value}
}
\description{
Efficient implementation of the Mann-Kendall test for trend detection using
merge sort algorithm (O(n log^2 n) complexity).
}
\examples{
 \dontrun{
data <- c(1.2, 3.4, 2.5, 4.1, 5.6)
result <- TH_MK_Trend(data)
}
}
\author{
Vladimiro Andrea Boselli, (2025) \email{boselli.v@irea.cnr.it}
}
