% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_plott.R
\name{TH_plott}
\alias{TH_plott}
\title{Plot hierarchical moving-window statistical analysis for time series.}
\usage{
TH_plott(
  results,
  scale = NULL,
  output_file = NULL,
  colorscales = NULL,
  colorlimits = NULL
)
}
\arguments{
\item{`results`}{is a \code{data.frame} created by TH_tweak().}

\item{`scale`}{a \code{numeric}, scaling factor for point sizes, autoset by default.}

\item{`output_file`}{is a \code{string}, optional, filename to save plot (NULL to skip saving).}

\item{`colorscales`}{a \code{list} of color scales for each function (optional). Can mix predefined names ("avg", "trend", "pval") and custom color vectors.}

\item{`colorlimits`}{a \code{list} of numeric vectors specifying color limits for each function (optional). Each element should be a vector of length 2 (c(min, max)). If NULL, limits are calculated automatically.}
}
\value{
A \code{gridExtra} arranged plot object.
}
\description{
Plots hierarchical moving-window statistical analysis for custom functions.
The input is a \code{data.frame} created by TH_tweak() function.
}
\examples{
 \dontrun{
results <- TH_tweak(fun1, fun2, fun3, series1, series2)
TH_plott(results)
TH_plott(results, colorscales = list(c("blue", "white", "red"), "avg", "trend"))
}
}
\author{
Vladimiro Andrea Boselli, (2025) \email{boselli.v@irea.cnr.it}
}
