% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_single.R
\name{TH_single}
\alias{TH_single}
\title{Perform hierarchical moving-window statistical analysis for time series.}
\usage{
TH_single(series, m = NULL, s = NULL, mode = "all", alpha = 0.1)
}
\arguments{
\item{`series`}{a \verb{numerical vector} containing NA for missing values.}

\item{`m`}{a \code{numeric} positive integer, sub-sampling parameter. Optional
(default: auto-calculated).}

\item{`s`}{a \code{numeric} positive integer, cutoff parameters for sub-series
length. Optional (default: auto-calculated).}

\item{`mode`}{a \code{character} string specifying the computation mode. One of:
"all" (default), "avg_only", "trend_only", "avg_trend", "trend_with_test", "avg_with_test".}

\item{`alpha`}{a \code{numeric} significance level for t-test flag (default: 0.1).}
}
\value{
A result \code{data.frame} to pass to TH_plots(), \code{s}, \code{m} and the length
of the series are passed as silent attributes.
}
\description{
Perform hierarchical moving-window statistical analysis for single time
series. It returns analysis on moving average and running trends. The results
are in a \code{data.frame} to be passed to TH_plots() function.
}
\examples{
 \dontrun{
output <- TH_single(
  series = rnorm(200),
  m = 2,
  s = 6,
  mode = "avg_with_test",
  alpha = 0.05
)
}
}
\author{
Vladimiro Andrea Boselli, (2025) \email{boselli.v@irea.cnr.it}
}
