% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_tweak.R
\name{TH_tweak}
\alias{TH_tweak}
\title{Perform hierarchical moving-window statistical analysis for time series.}
\usage{
TH_tweak(..., series, m = NULL, s = NULL, param = 0.1)
}
\arguments{
\item{...}{User-defined functions for analysis}

\item{series}{A \code{list} of \verb{numerical vectors} or a single \verb{numerical vector}}

\item{m}{a \code{numeric} positive integer, subsampling parameter. Optional (default: autocalculated).}

\item{s}{a \code{numeric} positive integer, cutoff parameters for subseries length. Optional (default: autocalculated).}

\item{param}{a \code{numeric} significance level for flagging (default: 0.1).}
}
\value{
A result \code{data.frame} to pass to TH_plott() with attributes.
}
\description{
Perform hierarchical moving-window statistical analysis using custom functions.
It returns analysis results in a \code{data.frame} to be passed to TH_plott() function.
}
\examples{
 \dontrun{
output <- TH_tweak(
  function(x, y) cor.test(x, y, method = "pearson")$estimate,
  function(x, y) cor.test(x, y, method = "kendall")$estimate,
  series = list(serie1, serie2)
)
}
}
\author{
Vladimiro Andrea Boselli, (2025) \email{boselli.v@irea.cnr.it}
}
