% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_MK_Corr.R
\name{TH_MK_Corr}
\alias{TH_MK_Corr}
\title{Mann-Kendall Tau Test with Knight's Algorithm}
\usage{
TH_MK_Corr(x, y, alternative = c("two.sided", "greater", "less"))
}
\arguments{
\item{x}{a \verb{numerical vector}}

\item{y}{a \verb{numerical vector}}

\item{alternative}{Type of alternative hypothesis ("two.sided", "greater", "less")}
}
\value{
A \code{htest} object with results
\item{statistic}{Kendall's tau-b statistic}
\item{p.value}{p-value}
\item{alternative}{Alternative hypothesis description}
}
\description{
Efficient implementation of the Mann-Kendall test for correlation using
Knight's algorithm of (O(n log n) complexity). Handles ties correctly.
}
\examples{
\dontrun{
kt <- TH_MK_Corr(sub_series1, sub_series2, alternative = "less")
c(kt$statistic, kt$p.value)
}
}
\author{
Vladimiro Andrea Boselli, (2025) \email{boselli.v@irea.cnr.it}
}
